+++ CHANGELOG.md
@@ -23,13 +23,13 @@ v1.0.0 (??.??.????)
  * new Wrapper Class Tx_Rnbase_Interface_Singleton (don't forget to load)  for \TYPO3\CMS\Core\SingletonInterface or t3lib_Singleton which is compatible for all TYPO3 versions
  * new Wrapper function tx_rnbase_util_Misc::logDeprecatedFunction
  * new Wrapper function tx_rnbase_util_Logger::devLog
- * new Wrapper function tx_rnbase_mod_Util::mapRecordTypeToSpriteIconName
  * new Wrapper function tx_rnbase_parameters::setGetParameter
  * new Wrapper function tx_rnbase_mod_BaseModule::issueCommand
  * new class tx_rnbase_util_Typo3Classes to get TYPO3 classes independent of TYPO3 version
  * new Wrapper Class Tx_Rnbase_CommandLine (don't forget to load)  for \TYPO3\CMS\Core\Controller\CommandLineController or t3lib_cli which is compatible for all TYPO3 versions
  * Bugfix: tx_rnbase_util_Dates::date_mysql2tstamp liefert nun wieder NULL im Fehlerfall
  * new Wrapper function tx_rnbase_util_Arrays::xml2array
+ * new Wrapper Class Tx_Rnbase_Backend_Utility_Icons (don't forget to load) for \TYPO3\CMS\Backend\Utility\IconUtility or t3lib_iconWorks which is compatible for all TYPO3 versions
 
 v0.15.18 (02.12.2015)
  * Bugfix: tx_rnbase_util_Dates::date_mysql2tstamp liefert nun immer einen Integer und nicht NULL im Fehlerfall